Add more links example site in Fig 1.
To crop an image in Inkscape, use the rectangle tool to cover the
part you want to keep. Then select both the image and the
rectangle, and do Object->Clip->Set.

http://goinkscape.com/how-to-crop-an-image-in-inkscape/

To add a border to an image in Inkscape, use the rectangle tool
to cover the image. Remove the fill and use a stroke of the
desired border color.

http://www.inkscapeforum.com/viewtopic.php?t=638